backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
kuzu
chromadb
sentence-transformers
gitpython
PyGithub
tree-sitter
tree-sitter-python
tree-sitter-go
tree-sitter-javascript
tree-sitter-typescript
tree-sitter-java
tree-sitter-rust
tree-sitter-c
tree-sitter-cpp
rank-bm25
groq
numpy
